code/requirements.txt
flask==2.2.3
selenium==4.8.2
beautifulsoup4==4.12.0
pyquery==2.0.0
pymysql==1.0.3
jieba==0.42.1
wordcloud==1.8.2.2
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
pillow==9.5.0
requests==2.28.2
tqdm==4.65.0
python-dotenv==1.0.0
flask-cors==3.0.10

